main/backend/capstone_project/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
# autobahn==21.3.1
Automat==20.2.0
beautifulsoup4==4.9.3
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
cognitive-face==1.5.0
constantly==15.1.0
cryptography==3.4.6
daphne==3.0.1
defusedxml==0.6.0
Django==3.1.6
django-allauth==0.44.0
django-cors-headers==3.7.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
geojson==2.5.0
hiredis==1.1.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
# mysqlclient==2.0.3
oauthlib==3.1.0
Pillow==8.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
# pyowm==3.2.0
PySocks==1.7.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
Twisted==21.2.0
txaio==21.2.1
urllib3==1.26.3
zope.interface==5.3.0
